perf(reward): backbone
