+++ contracts/proxyreward/src/reward.cpp
@@ -1,0 +1,10 @@
+#include <distribution.hpp>
+#include <reward.hpp>
+#include <voter.hpp>
+
+EOSIO_ACTION_DISPATCHER( edenproxy::actions )
+
+EOSIO_ABIGEN( actions( edenproxy::actions ),
+              table( "distribution"_n, edenproxy::distribution_variant ),
+              //   table( "settings"_n, edenproxy::settings_variant ),
+              table( "voter"_n, edenproxy::voter_variant ) )
